Fix wallet reconnect
