+++ frontend/src/components/ConnectWallet.tsx
@@ -1,6 +1,7 @@
 import { InjectedConnector, useStarknet } from "@starknet-react/core";
 import { useRouter } from "next/router";
-import { useEffect } from "react";
+import { useCallback, useEffect } from "react";
+import { useSnapshot } from "valtio";
 import Box from "~/components/Box";
 import Button from "~/components/Button";
 import { Symbol } from "~/components/Logos";
@@ -10,13 +11,20 @@ import { state } from "~/state";
 export function ConnectWallet() {
   const router = useRouter()
   const { connect, account } = useStarknet()
-
+  const { walletAddress } = useSnapshot(state)
+  
   useEffect(() => {
-    if (account) {
+    if (account && account !== walletAddress) {
+      console.log("EBINEINBIEN")
       state.walletAddress = account
       router.push("/create")
     }
-  }, [account])
+  }, [account, router, walletAddress])
+
+  const connectings = useCallback(() => {
+    console.log("ebinclickings")
+    connect(new InjectedConnector())
+  }, [connect])
 
   return (
     <>
@@ -34,7 +42,7 @@ export function ConnectWallet() {
         Get started by connecting your wallet. This allows you to create new
         multisignature contracts or use existing contracts.
       </Paragraph>
-      <Button fullWidth onClick={() => connect(new InjectedConnector())}>
+      <Button fullWidth onClick={connectings}>
         Connect wallet (Argent X)
       </Button>
     </>
